+++ WatchWebShop/Models/Product.cs
@@ -9,7 +9,7 @@ namespace WatchWebShop.Models
         public int Id { get; set; }
 
         public string Name { get; set; }
-        public decimal UnitPriceNetto { get; set; }
+        public double UnitPriceNetto { get; set; }
         public string ImagePath { get; set; }
         public string Description { get; set; }
